Added Dual Investment

+++ contracts/contracts/FeedsPriceOracle.sol
@@ -7,10 +7,7 @@ import "./PErc20.sol";
 interface IFeeds {
     function latestAnswer() external view returns (uint256);
 
-    function latestRoundData()
-        external
-        view
-        returns (uint80, uint256, uint256, uint256, uint80);
+    function latestRoundData() external view returns (uint80, uint256, uint256, uint256, uint80);
 
     function decimals() external view returns (uint8);
 
@@ -32,11 +29,7 @@ contract FeedsAccess {
         return feeds.latestAnswer();
     }
 
-    function fetchLatestRoundData()
-        external
-        view
-        returns (uint80, uint256, uint256, uint256, uint80)
-    {
+    function fetchLatestRoundData() external view returns (uint80, uint256, uint256, uint256, uint80) {
         return feeds.latestRoundData();
     }
 
@@ -55,23 +48,20 @@ contract FeedsAccess {
  * @author Peridot
  */
 contract FeedsPriceOracle is PriceOracle {
-    mapping(address => uint) prices;
+    mapping(address => uint256) prices;
     mapping(address => bool) public admin;
     mapping(address => address) public assetToFeedsAccess; // Maps asset addresses to FeedsAccess contract addresses
-    mapping(address => uint) public feedsStaleThreshold; // Maximum age of price feed in seconds for each asset
-    mapping(address => uint) public lastValidFeedsPriceMantissa; // Stores the last valid price mantissa from Feeds
+    mapping(address => uint256) public feedsStaleThreshold; // Maximum age of price feed in seconds for each asset
+    mapping(address => uint256) public lastValidFeedsPriceMantissa; // Stores the last valid price mantissa from Feeds
     address private owner;
-    uint public defaultStaleThreshold; // Default stale threshold for feeds
+    uint256 public defaultStaleThreshold; // Default stale threshold for feeds
 
     event PricePosted(
-        address asset,
-        uint previousPriceMantissa,
-        uint requestedPriceMantissa,
-        uint newPriceMantissa
+        address asset, uint256 previousPriceMantissa, uint256 requestedPriceMantissa, uint256 newPriceMantissa
     );
     event FeedRegistered(address asset, address feedsAccess);
-    event LastFeedsPriceUpdated(address indexed asset, uint priceMantissa);
-    event StaleThresholdUpdated(address indexed asset, uint newThreshold);
+    event LastFeedsPriceUpdated(address indexed asset, uint256 priceMantissa);
+    event StaleThresholdUpdated(address indexed asset, uint256 newThreshold);
 
     modifier onlyAdmin() {
         require(admin[msg.sender], "Only admin can call this function");
@@ -83,15 +73,13 @@ contract FeedsPriceOracle is PriceOracle {
         _;
     }
 
-    constructor(uint _defaultStaleThreshold) {
+    constructor(uint256 _defaultStaleThreshold) {
         owner = msg.sender;
         admin[msg.sender] = true;
         defaultStaleThreshold = _defaultStaleThreshold;
     }
 
-    function _getUnderlyingAddress(
-        PToken pToken
-    ) private view returns (address) {
+    function _getUnderlyingAddress(PToken pToken) private view returns (address) {
         address asset;
         if (compareStrings(pToken.symbol(), "pETH")) {
             asset = 0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE;
@@ -101,22 +89,18 @@ contract FeedsPriceOracle is PriceOracle {
         return asset;
     }
 
-    function getUnderlyingPrice(
-        PToken pToken
-    ) external view override returns (uint) {
+    function getUnderlyingPrice(PToken pToken) external view override returns (uint256) {
         address asset = _getUnderlyingAddress(pToken);
         address feedsAccessAddress = assetToFeedsAccess[asset];
 
         if (feedsAccessAddress != address(0)) {
-            try this._getFeedsPrice(feedsAccessAddress, asset) returns (
-                uint priceMantissa
-            ) {
+            try this._getFeedsPrice(feedsAccessAddress, asset) returns (uint256 priceMantissa) {
                 if (priceMantissa > 0) {
                     return priceMantissa;
                 }
             } catch {
                 // If feeds call fails, try returning last known valid price
-                uint lastValidPrice = lastValidFeedsPriceMantissa[asset];
+                uint256 lastValidPrice = lastValidFeedsPriceMantissa[asset];
                 if (lastValidPrice != 0) {
                     return lastValidPrice;
                 }
@@ -125,7 +109,7 @@ contract FeedsPriceOracle is PriceOracle {
         }
 
         // Fallback to manually set price
-        uint manualPrice = prices[asset];
+        uint256 manualPrice = prices[asset];
         if (manualPrice > 0) {
             return manualPrice;
         }
@@ -140,40 +124,29 @@ contract FeedsPriceOracle is PriceOracle {
      * @param asset The asset address for stale threshold lookup
      * @return The price mantissa scaled to 18 decimals
      */
-    function _getFeedsPrice(
-        address feedsAccessAddress,
-        address asset
-    ) external view returns (uint) {
+    function _getFeedsPrice(address feedsAccessAddress, address asset) external view returns (uint256) {
         require(msg.sender == address(this), "Internal function only");
 
         FeedsAccess feedsAccess = FeedsAccess(feedsAccessAddress);
 
         // Get the latest round data which includes timestamp
-        (
-            uint80 roundId,
-            uint256 price,
-            uint256 startedAt,
-            uint256 updatedAt,
-            uint80 answeredInRound
-        ) = feedsAccess.fetchLatestRoundData();
+        (uint80 roundId, uint256 price, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound) =
+            feedsAccess.fetchLatestRoundData();
 
         // Check if the price is stale
-        uint staleThreshold = feedsStaleThreshold[asset];
+        uint256 staleThreshold = feedsStaleThreshold[asset];
         if (staleThreshold == 0) {
             staleThreshold = defaultStaleThreshold;
         }
 
-        require(
-            block.timestamp - updatedAt <= staleThreshold,
-            "Price feed is stale"
-        );
+        require(block.timestamp - updatedAt <= staleThreshold, "Price feed is stale");
         require(price > 0, "Invalid price from feed");
 
         // Get decimals from the feed
         uint8 feedDecimals = feedsAccess.getDecimals();
 
         // Convert price to 18 decimals
-        uint priceMantissa = price;
+        uint256 priceMantissa = price;
         if (feedDecimals < 18) {
             priceMantissa = priceMantissa * (10 ** (18 - feedDecimals));
         } else if (feedDecimals > 18) {
@@ -189,15 +162,10 @@ contract FeedsPriceOracle is PriceOracle {
      */
     function updateCachedPrice(address asset) external {
         address feedsAccessAddress = assetToFeedsAccess[asset];
-        require(
-            feedsAccessAddress != address(0),
-            "No feed registered for asset"
-        );
+        require(feedsAccessAddress != address(0), "No feed registered for asset");
 
-        try this._getFeedsPrice(feedsAccessAddress, asset) returns (
-            uint priceMantissa
-        ) {
-            uint oldPrice = lastValidFeedsPriceMantissa[asset];
+        try this._getFeedsPrice(feedsAccessAddress, asset) returns (uint256 priceMantissa) {
+            uint256 oldPrice = lastValidFeedsPriceMantissa[asset];
             lastValidFeedsPriceMantissa[asset] = priceMantissa;
             emit LastFeedsPriceUpdated(asset, priceMantissa);
         } catch {
@@ -211,17 +179,9 @@ contract FeedsPriceOracle is PriceOracle {
      * @param pToken The pToken to set the price for
      * @param underlyingPriceMantissa The price in 18 decimal format
      */
-    function setUnderlyingPrice(
-        PToken pToken,
-        uint underlyingPriceMantissa
-    ) public onlyAdmin {
+    function setUnderlyingPrice(PToken pToken, uint256 underlyingPriceMantissa) public onlyAdmin {
         address asset = _getUnderlyingAddress(pToken);
-        emit PricePosted(
-            asset,
-            prices[asset],
-            underlyingPriceMantissa,
-            underlyingPriceMantissa
-        );
+        emit PricePosted(asset, prices[asset], underlyingPriceMantissa, underlyingPriceMantissa);
         prices[asset] = underlyingPriceMantissa;
     }
 
@@ -230,7 +190,7 @@ contract FeedsPriceOracle is PriceOracle {
      * @param asset The asset address
      * @param price The price in 18 decimal format
      */
-    function setDirectPrice(address asset, uint price) public onlyAdmin {
+    function setDirectPrice(address asset, uint256 price) public onlyAdmin {
         emit PricePosted(asset, prices[asset], price, price);
         prices[asset] = price;
     }
@@ -240,10 +200,7 @@ contract FeedsPriceOracle is PriceOracle {
      * @param asset The asset address
      * @param feedsAccessAddress The FeedsAccess contract address (can be address(0) to unregister)
      */
-    function registerFeed(
-        address asset,
-        address feedsAccessAddress
-    ) public onlyAdmin {
+    function registerFeed(address asset, address feedsAccessAddress) public onlyAdmin {
         // Allow address(0) to unregister a feed
         assetToFeedsAccess[asset] = feedsAccessAddress;
         emit FeedRegistered(asset, feedsAccessAddress);
@@ -254,10 +211,7 @@ contract FeedsPriceOracle is PriceOracle {
      * @param asset The asset address
      * @param threshold The stale threshold in seconds
      */
-    function setAssetStaleThreshold(
-        address asset,
-        uint threshold
-    ) public onlyAdmin {
+    function setAssetStaleThreshold(address asset, uint256 threshold) public onlyAdmin {
         feedsStaleThreshold[asset] = threshold;
         emit StaleThresholdUpdated(asset, threshold);
     }
@@ -266,7 +220,7 @@ contract FeedsPriceOracle is PriceOracle {
      * @notice Set the default stale threshold for all assets
      * @param threshold The default stale threshold in seconds
      */
-    function setDefaultStaleThreshold(uint threshold) public onlyOwner {
+    function setDefaultStaleThreshold(uint256 threshold) public onlyOwner {
         defaultStaleThreshold = threshold;
     }
 
@@ -299,19 +253,17 @@ contract FeedsPriceOracle is PriceOracle {
      * @param asset The asset address
      * @return The price in 18 decimal format
      */
-    function assetPrices(address asset) external view returns (uint) {
+    function assetPrices(address asset) external view returns (uint256) {
         address feedsAccessAddress = assetToFeedsAccess[asset];
 
         if (feedsAccessAddress != address(0)) {
-            try this._getFeedsPrice(feedsAccessAddress, asset) returns (
-                uint priceMantissa
-            ) {
+            try this._getFeedsPrice(feedsAccessAddress, asset) returns (uint256 priceMantissa) {
                 if (priceMantissa > 0) {
                     return priceMantissa;
                 }
             } catch {
                 // If feeds call fails, try returning last known valid price
-                uint lastValidPrice = lastValidFeedsPriceMantissa[asset];
+                uint256 lastValidPrice = lastValidFeedsPriceMantissa[asset];
                 if (lastValidPrice != 0) {
                     return lastValidPrice;
                 }
@@ -320,7 +272,7 @@ contract FeedsPriceOracle is PriceOracle {
         }
 
         // Fallback to manually set price
-        uint manualPrice = prices[asset];
+        uint256 manualPrice = prices[asset];
         if (manualPrice > 0) {
             return manualPrice;
         }
@@ -334,13 +286,8 @@ contract FeedsPriceOracle is PriceOracle {
      * @param feedAggregator The address of the feed aggregator
      * @return The address of the deployed FeedsAccess contract
      */
-    function deployFeedsAccess(
-        address feedAggregator
-    ) external onlyAdmin returns (address) {
-        require(
-            feedAggregator != address(0),
-            "Invalid feed aggregator address"
-        );
+    function deployFeedsAccess(address feedAggregator) external onlyAdmin returns (address) {
+        require(feedAggregator != address(0), "Invalid feed aggregator address");
 
         FeedsAccess feedsAccess = new FeedsAccess(feedAggregator);
         return address(feedsAccess);
@@ -360,11 +307,7 @@ contract FeedsPriceOracle is PriceOracle {
      * @param b Second string
      * @return True if strings are equal
      */
-    function compareStrings(
-        string memory a,
-        string memory b
-    ) internal pure returns (bool) {
-        return (keccak256(abi.encodePacked((a))) ==
-            keccak256(abi.encodePacked((b))));
+    function compareStrings(string memory a, string memory b) internal pure returns (bool) {
+        return (keccak256(abi.encodePacked((a))) == keccak256(abi.encodePacked((b))));
     }
 }
