+++ contracts/contracts/JumpRateModel.sol
@@ -9,10 +9,7 @@ import "./InterestRateModel.sol";
  */
 contract JumpRateModel is InterestRateModel {
     event NewInterestParams(
-        uint baseRatePerBlock,
-        uint multiplierPerBlock,
-        uint jumpMultiplierPerBlock,
-        uint kink
+        uint256 baseRatePerBlock, uint256 multiplierPerBlock, uint256 jumpMultiplierPerBlock, uint256 kink
     );
 
     uint256 private constant BASE = 1e18;
@@ -20,27 +17,27 @@ contract JumpRateModel is InterestRateModel {
     /**
      * @notice The approximate number of blocks per year that is assumed by the interest rate model
      */
-    uint public constant blocksPerYear = 2102400;
+    uint256 public constant blocksPerYear = 2102400;
 
     /**
      * @notice The multiplier of utilization rate that gives the slope of the interest rate
      */
-    uint public multiplierPerBlock;
+    uint256 public multiplierPerBlock;
 
     /**
      * @notice The base interest rate which is the y-intercept when utilization rate is 0
      */
-    uint public baseRatePerBlock;
+    uint256 public baseRatePerBlock;
 
     /**
      * @notice The multiplierPerBlock after hitting a specified utilization point
      */
-    uint public jumpMultiplierPerBlock;
+    uint256 public jumpMultiplierPerBlock;
 
     /**
      * @notice The utilization point at which the jump multiplier is applied
      */
-    uint public kink;
+    uint256 public kink;
 
     /**
      * @notice Construct an interest rate model
@@ -49,23 +46,13 @@ contract JumpRateModel is InterestRateModel {
      * @param jumpMultiplierPerYear The multiplierPerBlock after hitting a specified utilization point
      * @param kink_ The utilization point at which the jump multiplier is applied
      */
-    constructor(
-        uint baseRatePerYear,
-        uint multiplierPerYear,
-        uint jumpMultiplierPerYear,
-        uint kink_
-    ) {
+    constructor(uint256 baseRatePerYear, uint256 multiplierPerYear, uint256 jumpMultiplierPerYear, uint256 kink_) {
         baseRatePerBlock = baseRatePerYear / blocksPerYear;
         multiplierPerBlock = multiplierPerYear / blocksPerYear;
         jumpMultiplierPerBlock = jumpMultiplierPerYear / blocksPerYear;
         kink = kink_;
 
-        emit NewInterestParams(
-            baseRatePerBlock,
-            multiplierPerBlock,
-            jumpMultiplierPerBlock,
-            kink
-        );
+        emit NewInterestParams(baseRatePerBlock, multiplierPerBlock, jumpMultiplierPerBlock, kink);
     }
 
     /**
@@ -75,11 +62,7 @@ contract JumpRateModel is InterestRateModel {
      * @param reserves The amount of reserves in the market (currently unused)
      * @return The utilization rate as a mantissa between [0, BASE]
      */
-    function utilizationRate(
-        uint cash,
-        uint borrows,
-        uint reserves
-    ) public pure returns (uint) {
+    function utilizationRate(uint256 cash, uint256 borrows, uint256 reserves) public pure returns (uint256) {
         // Utilization rate is 0 when there are no borrows
         if (borrows == 0) {
             return 0;
@@ -95,19 +78,14 @@ contract JumpRateModel is InterestRateModel {
      * @param reserves The amount of reserves in the market
      * @return The borrow rate percentage per block as a mantissa (scaled by BASE)
      */
-    function getBorrowRate(
-        uint cash,
-        uint borrows,
-        uint reserves
-    ) public view override returns (uint) {
-        uint util = utilizationRate(cash, borrows, reserves);
+    function getBorrowRate(uint256 cash, uint256 borrows, uint256 reserves) public view override returns (uint256) {
+        uint256 util = utilizationRate(cash, borrows, reserves);
 
         if (util <= kink) {
             return ((util * multiplierPerBlock) / BASE) + baseRatePerBlock;
         } else {
-            uint normalRate = ((kink * multiplierPerBlock) / BASE) +
-                baseRatePerBlock;
-            uint excessUtil = util - kink;
+            uint256 normalRate = ((kink * multiplierPerBlock) / BASE) + baseRatePerBlock;
+            uint256 excessUtil = util - kink;
             return ((excessUtil * jumpMultiplierPerBlock) / BASE) + normalRate;
         }
     }
@@ -120,15 +98,15 @@ contract JumpRateModel is InterestRateModel {
      * @param reserveFactorMantissa The current reserve factor for the market
      * @return The supply rate percentage per block as a mantissa (scaled by BASE)
      */
-    function getSupplyRate(
-        uint cash,
-        uint borrows,
-        uint reserves,
-        uint reserveFactorMantissa
-    ) public view override returns (uint) {
-        uint oneMinusReserveFactor = BASE - reserveFactorMantissa;
-        uint borrowRate = getBorrowRate(cash, borrows, reserves);
-        uint rateToPool = (borrowRate * oneMinusReserveFactor) / BASE;
+    function getSupplyRate(uint256 cash, uint256 borrows, uint256 reserves, uint256 reserveFactorMantissa)
+        public
+        view
+        override
+        returns (uint256)
+    {
+        uint256 oneMinusReserveFactor = BASE - reserveFactorMantissa;
+        uint256 borrowRate = getBorrowRate(cash, borrows, reserves);
+        uint256 rateToPool = (borrowRate * oneMinusReserveFactor) / BASE;
         return (utilizationRate(cash, borrows, reserves) * rateToPool) / BASE;
     }
 }
