+++ contracts/contracts/Maximillion.sol
@@ -35,12 +35,9 @@ contract Maximillion {
      * @param borrower The address of the borrower account to repay on behalf of
      * @param cEther_ The address of the pEther contract to repay in
      */
-    function repayBehalfExplicit(
-        address borrower,
-        PEther cEther_
-    ) public payable {
-        uint received = msg.value;
-        uint borrows = cEther_.borrowBalanceCurrent(borrower);
+    function repayBehalfExplicit(address borrower, PEther cEther_) public payable {
+        uint256 received = msg.value;
+        uint256 borrows = cEther_.borrowBalanceCurrent(borrower);
         if (received > borrows) {
             cEther_.repayBorrowBehalf{value: borrows}(borrower);
             payable(msg.sender).transfer(received - borrows);
