+++ contracts/contracts/MockOracle.sol
@@ -5,17 +5,13 @@ import "./PriceOracle.sol";
 import "./PErc20.sol";
 
 contract MockOracle is PriceOracle {
-    mapping(address => uint) prices;
+    mapping(address => uint256) prices;
+
     event PricePosted(
-        address asset,
-        uint previousPriceMantissa,
-        uint requestedPriceMantissa,
-        uint newPriceMantissa
+        address asset, uint256 previousPriceMantissa, uint256 requestedPriceMantissa, uint256 newPriceMantissa
     );
 
-    function _getUnderlyingAddress(
-        PToken pToken
-    ) private view returns (address) {
+    function _getUnderlyingAddress(PToken pToken) private view returns (address) {
         address asset;
         if (compareStrings(pToken.symbol(), "pETH")) {
             asset = 0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE;
@@ -25,41 +21,27 @@ contract MockOracle is PriceOracle {
         return asset;
     }
 
-    function getUnderlyingPrice(
-        PToken pToken
-    ) public view override returns (uint) {
+    function getUnderlyingPrice(PToken pToken) public view override returns (uint256) {
         return prices[_getUnderlyingAddress(pToken)];
     }
 
-    function setUnderlyingPrice(
-        PToken pToken,
-        uint underlyingPriceMantissa
-    ) public {
+    function setUnderlyingPrice(PToken pToken, uint256 underlyingPriceMantissa) public {
         address asset = _getUnderlyingAddress(pToken);
-        emit PricePosted(
-            asset,
-            prices[asset],
-            underlyingPriceMantissa,
-            underlyingPriceMantissa
-        );
+        emit PricePosted(asset, prices[asset], underlyingPriceMantissa, underlyingPriceMantissa);
         prices[asset] = underlyingPriceMantissa;
     }
 
-    function setDirectPrice(address asset, uint price) public {
+    function setDirectPrice(address asset, uint256 price) public {
         emit PricePosted(asset, prices[asset], price, price);
         prices[asset] = price;
     }
 
     // v1 price oracle interface for use as backing of proxy
-    function assetPrices(address asset) external view returns (uint) {
+    function assetPrices(address asset) external view returns (uint256) {
         return prices[asset];
     }
 
-    function compareStrings(
-        string memory a,
-        string memory b
-    ) internal pure returns (bool) {
-        return (keccak256(abi.encodePacked((a))) ==
-            keccak256(abi.encodePacked((b))));
+    function compareStrings(string memory a, string memory b) internal pure returns (bool) {
+        return (keccak256(abi.encodePacked((a))) == keccak256(abi.encodePacked((b))));
     }
 }
