+++ contracts/contracts/PriceOracle.sol
@@ -13,7 +13,5 @@ abstract contract PriceOracle {
      * @return The underlying asset price mantissa (scaled by 1e18).
      *  Zero means the price is unavailable.
      */
-    function getUnderlyingPrice(
-        PToken pToken
-    ) external view virtual returns (uint);
+    function getUnderlyingPrice(PToken pToken) external view virtual returns (uint256);
 }
